# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TechAtlas API
  description: ''
  contact:
    name: TechAtlas Contributors
  license:
    name: MIT
    identifier: MIT
  version: 1.0.0
paths:
...
  /api/v1/admin/country-enrichment/recrawl:
    post:
      tags:
      - admin-operations
      operationId: schedule_country_enrichment_recrawl
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryEnrichmentRecrawlResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
  /api/v1/admin/crawl-attempts:
    get:
      tags:
      - admin-operations
      operationId: crawl_attempts
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlAttemptPageResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
...
  /api/v1/admin/domains:
    get:
      tags:
      - admin-domains
      operationId: list_domains
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Page of active domains
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainListResponse'
        '401':
          description: Authentication failed
          headers:
            WWW-Authenticate:
              schema:
                type: string
              description: Bearer authentication challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Query validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds before retrying
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
    post:
      tags:
      - admin-domains
      operationId: create_domain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomainRequest'
        required: true
      responses:
        '201':
          description: Domain created
          headers:
            Location:
              schema:
                type: string
              description: Canonical resource location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainResponse'
        '401':
          description: Authentication failed
          headers:
            WWW-Authenticate:
              schema:
                type: string
              description: Bearer authentication challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Domain already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds before retrying
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
...
  /api/v1/admin/domains/{canonical_domain}/crawl:
    post:
      tags:
      - admin-domains
      operationId: request_domain_crawl
      parameters:
      - name: canonical_domain
        in: path
        description: Existing enabled domain name or normalizable HTTP(S) URL
        required: true
        schema:
          type: string
      responses:
        '202':
          description: Domain made eligible for scheduler processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrawlRequestResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Crawl policy is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
...
  /api/v1/admin/operations/overview:
    get:
      tags:
      - admin-operations
      operationId: operations_overview
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsOverviewResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - oidc_bearer: []
...
components:
  schemas:
...
    AdminCrawlRetryState:
      type: string
      enum:
      - in_progress
      - automatic_retry_scheduled
      - manual_retry_available
      - manual_retry_blocked
      - not_applicable
...
    CountryEnrichmentRecrawlResponse:
      type: object
      required:
      - scheduled_domain_count
      properties:
        scheduled_domain_count:
          type: integer
          format: int64
          minimum: 0
    CrawlAttemptPageResponse:
      type: object
      required:
      - attempts
      properties:
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/CrawlAttemptResponse'
    CrawlAttemptResponse:
      type: object
      required:
      - job_id
      - canonical_domain
      - attempt_number
      - status
      - queued_at
      - retry_state
      - retry_eligible
      properties:
        attempt_number:
          type: integer
          format: int32
          minimum: 0
        canonical_domain:
          type: string
        failure_code:
          type:
          - string
          - 'null'
        failure_summary:
          type:
          - string
          - 'null'
        finished_at:
          type:
          - string
          - 'null'
        job_id:
          type: string
        queued_at:
          type: string
        retry_eligible:
          type: boolean
        retry_state:
          $ref: '#/components/schemas/AdminCrawlRetryState'
        status:
          type: string
    CrawlRequestResponse:
      type: object
      required:
      - canonical_domain
      - scheduled_at
      properties:
        canonical_domain:
          type: string
        scheduled_at:
          type: string
    CreateDomainRequest:
      type: object
      required:
      - domain
      properties:
        domain:
          type: string
...
    DependencyStatusResponse:
      type: object
      required:
      - name
      - status
      properties:
        name:
          type: string
        status:
          type: string
...
    DomainListResponse:
      type: object
      required:
      - domains
      properties:
        domains:
          type: array
          items:
            $ref: '#/components/schemas/DomainResponse'
        next_cursor:
          type:
          - string
          - 'null'
    DomainResponse:
      type: object
      required:
      - id
      - canonical_domain
      properties:
        canonical_domain:
          type: string
        id:
          type: string
    ErrorBody:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    OperationsActivityResponse:
      type: object
      required:
      - id
      - title
      - description
      - occurred_at
      - kind
      properties:
        description:
          type: string
        id:
          type: string
        kind:
          type: string
        occurred_at:
          type: string
        title:
          type: string
    OperationsAlertResponse:
      type: object
      required:
      - code
      - severity
      - message
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
    OperationsOverviewResponse:
      type: object
      required:
      - generated_at
      - system_status
      - domain_count
      - technology_count
      - current_detection_count
      - successful_crawl_count
      - queue
      - dependencies
      - alertable_failures
      - throughput
      - workers
      - activity
      properties:
        activity:
          type: array
          items:
            $ref: '#/components/schemas/OperationsActivityResponse'
        alertable_failures:
          type: array
          items:
            $ref: '#/components/schemas/OperationsAlertResponse'
        current_detection_count:
          type: integer
          format: int64
          minimum: 0
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/DependencyStatusResponse'
        domain_count:
          type: integer
          format: int64
          minimum: 0
        generated_at:
          type: string
        queue:
          $ref: '#/components/schemas/QueueOverviewResponse'
        successful_crawl_count:
          type: integer
          format: int64
          minimum: 0
        system_status:
          type: string
        technology_count:
          type: integer
          format: int64
          minimum: 0
        throughput:
          type: array
          items:
            $ref: '#/components/schemas/ThroughputPointResponse'
        workers:
          type: array
          items:
            $ref: '#/components/schemas/WorkerOverviewResponse'
...
    QueueOverviewResponse:
      type: object
      required:
      - ready_count
      - processing_count
      - scheduled_count
      properties:
        processing_count:
          type: integer
          format: int64
          minimum: 0
        ready_count:
          type: integer
          format: int64
          minimum: 0
        scheduled_count:
          type: integer
          format: int64
          minimum: 0
...
    ThroughputPointResponse:
      type: object
      required:
      - observed_at
      - completed_count
      properties:
        completed_count:
          type: integer
          format: int64
          minimum: 0
        observed_at:
          type: string
...
    WorkerOverviewResponse:
      type: object
      required:
      - name
      - region
      - in_flight_work
      - completed_total
      - status
      - last_heartbeat_at
      properties:
        completed_total:
          type: integer
          format: int64
          minimum: 0
        in_flight_work:
          type: integer
          format: int32
          minimum: 0
        last_heartbeat_at:
          type: string
        name:
          type: string
        region:
          type: string
        status:
          type: string
  securitySchemes:
    oidc_bearer:
      type: http
      scheme: bearer
      bearerFormat: Bearer